week-3/requirements.txt
tensorflow>=2.10.0
scikit-learn>=1.1.0
matplotlib>=3.6.0
streamlit>=1.15.0
Pillow>=9.3.0 